requirements.txt
python-telegram-bot==20.6
selenium==4.13.0
webdriver-manager==4.0.1
asyncio==3.4.3